Make next-server work standalone
