+++ packages/next-server/package.json
@@ -24,10 +24,12 @@
     ]
   },
   "dependencies": {
+    "@babel/runtime-corejs2": "^7.1.2",
     "ansi-html": "0.0.7",
     "etag": "1.8.1",
     "find-up": "3.0.0",
     "fresh": "0.5.2",
+    "hoist-non-react-statics": "^3.0.1",
     "htmlescape": "1.1.1",
     "http-errors": "1.6.2",
     "path-to-regexp": "2.1.0",
