+++ packages/next/error.js
@@ -1,1 +1,1 @@
-module.exports = require('./dist/lib/error')
+module.exports = require('./dist/pages/_error')
